backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
apscheduler==3.10.4
python-multipart==0.0.6
msal==1.26.0
google-auth==2.26.1
google-api-python-client==2.115.0
cryptography==41.0.7
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
greenlet==3.0.3
psycopg2-binary==2.9.9
aiohttp==3.9.1
